% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_plot_RR.R
\name{Define_RR}
\alias{Define_RR}
\title{Define_RR: Defines zones for plotting Rogers Ratio Diagnostic Logic as per IEEE Table 5}
\usage{
Define_RR()
}
\value{
Data.frame:
\itemize{
\item \code{xmin} | \code{xmax} | \code{ymin} | \code{ymax}: Numeric. Used to define rectangles in a ggplot.
\item \code{label}: String. Used for colour scaling in ggplot.
\item \code{name}: String. Used for defining which facet to plot in.
}
}
\description{
This defines zones for plotting Rogers Ratio in accordance to IEEE C57.104-2019 Table 5.
}
