% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_logic_DTn.R
\name{Interpret_DTn}
\alias{Interpret_DTn}
\title{Interpret_DTn: Applies Duval Triangles 1, 4, or 5 Diagnostic Logic as per IEEE Table 6, D.3, and D.4, respectively.}
\usage{
Interpret_DTn(DT_n, y_in, n_in)
}
\arguments{
\item{DT_n}{String: Value corresponding to either Triangle 1, 4, or 5. Expected value: 1, 4, or 5.}

\item{y_in}{Numeric Array: Values representing generated gas ratios, assumed order: DT1: C2H4, CH4, C2H2, DT4: CH4, H2, C2H6, DT5: C2H4, CH4, C2H6. Assumed shape: 3 x Samples.}

\item{n_in}{Numeric: Number of samples to be diagnosed.}
}
\value{
Strings: List of Diagnoses, one for each inputted sample.
}
\description{
This functions interprets the inputted gas sample ratios to determine the appropriate diagnosis for each in accordance to either IEEE C57.104-2019 Table 6 or Table D.3, or Table D.4. Table D.3 is modified here such that original definition for C: (CH4 >= 36 & C2H6 >= 24) is instead: (CH4 >= 36 & C2H6 < 24).
}
\examples{
Interpret_DTn(DT_n = "4",
               y_in = array(c(0.090, 0.002, 0.401, 0.120, 0.001, 0.310), dim = c(3,2)),
               n_in = 2)
}
