% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_logic_RR.R
\name{Interpret_RR}
\alias{Interpret_RR}
\title{Interpret_RR: Applies Rogers Ratio Diagnostic Logic as per IEEE Table 5}
\usage{
Interpret_RR(y_in, n_in)
}
\arguments{
\item{y_in}{Numeric Array: Values representing generated gas ratios, assumed order: C2H2/C2H4, CH4/H2, C2H4/C2H6, assumed shape: 3 x Samples.}

\item{n_in}{Numeric: Number of samples to be diagnosed.}
}
\value{
Strings: List of Diagnoses, one for each inputted sample.
}
\description{
This functions interprets the inputted gas sample ratios to determine the appropriate diagnosis for each in accordance to IEEE C57.104-2019 Table 5.
}
\examples{
Interpret_RR(y_in = array(c(0.090, 0.002, 0.401, 0.120, 0.001, 0.310), dim = c(3,2)),
             n_in = 2)
}
