% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{Preprocess}
\alias{Preprocess}
\title{Preprocess: Validate inputs and preprocess them into expected data formats / values}
\usage{
Preprocess(
  gas_data,
  limit_data,
  accuracy,
  min_accuracy,
  distribution_shape,
  rho,
  diagnosis,
  n,
  flex = FALSE,
  silent = FALSE,
  verbose = FALSE,
  validate = FALSE
)
}
\arguments{
\item{gas_data}{Data.frame: First column containing date of sample. Cannot have two samples on same day. Rest of columns containing sample value in PPM.}

\item{limit_data}{Data.frame: Columns containing the IEEE DGA Table limits for each gas. Each Table is on its own row; four rows expected.}

\item{accuracy}{Named Numeric: Values to be multiplied by gas value to give uncertainty. Can either be same value for all gases, or individual values for each gas. Expected range: 0-1.}

\item{min_accuracy}{Named Numeric: Absolute Values (PPM) used as minimum uncertainty. Can either be same value for all gases, or individual values for each gas.}

\item{distribution_shape}{Numeric: Distribution shape representing uncertainty. "N" = Normal, "T" = Triangular, "U" = Uniform. If Normal, Accuracy is interpreted as 95\% confidence interval, else as 100\%.}

\item{rho}{Numeric: Value representing rho for correlations. Can either be same value for all gases, individual values for each gas, or a matrix intended to represent the \code{cor_mat}.}

\item{diagnosis}{Mixed: Used to specify which diagnoses to conduct. 0 = None, 1 = All, "DT" = Duval Triangles 1|4|5 only, "RR" = Rogers Ratio Only, "IEC" = IEC Tables 1|2 Only.}

\item{n}{Numeric: Number of samples to generate for the Monte Carlo for each gas. If a value <1, it is assumed as an estimate for the error tolerance, then Wald Interval used to estimate n.}

\item{flex}{Boolean: Defaulted to FALSE. If TRUE, and \code{rho} is assumed a correlation matrix equivalent, it will attempt to force it to become a positive definite if needed.}

\item{silent}{Boolean: Defaulted to FALSE.}

\item{verbose}{Boolean: Defaulted to FALSE.}

\item{validate}{Boolean: Defaulted to FALSE.}
}
\value{
List containing:
\itemize{
\item $\code{gas_data}: Data.frame. First column as Date converted to days since oldest sample. Rest as gas values in PPM.
\item $\code{limit_data}: Data.frame. Unchanged from input.
\item $\code{accuracy}: Numeric. Unchanged from input.
\item $\code{min_accuracy}: Numeric. Unchanged from input.
\item $\code{distribution_shape}: Numeric. 1 = Normal, 2 = Triangular, 3 = Uniform.
\item $\code{rho}: Numeric or NA. NA if \code{cor_mat} is being used.
\item $\code{diagnosis}: Numeric. 0 = None, 1 = All, 2 = Duval Triangles, 3 = Rogers Ratios, 4 = IEC Tables.
\item $\code{n}: Numeric. Number of samples used for Monte Carlo Simulation.
\item $\code{col_names}: Names of found gas names.
\item $\code{cor_mat}: Matrix or NA. NA if \code{rho} is being used.
\item $\code{sd_data}: Data.frame. Same as \code{gas_data} but with the values representing standard deviation and not gas values.
}
}
\description{
This function is intended to capture errors. No further validation is then done beyond this function.
}
\examples{
Preprocess(gas_data = data.frame(date = as.Date(c("2024-06-20","2024-06-21","2024-06-22","2024-06-23","2024-06-24")), H2 = c(8,4,2,1,0.5), CH4 = c(10,20,40,80,160)),
           limit_data = data.frame(H2 = c(2, 3, 2, 0.05), CH4 = c(60,180,70,0.03)),
           accuracy = setNames(c(0.15, 0.50), c("H2", "CH4")),
           min_accuracy = setNames(c(1, 5), c("H2", "CH4")),
           distribution_shape = "N",
           rho = 0.5,
           diagnosis = 0,
           n = 1000,
           flex = FALSE, silent = FALSE, verbose = FALSE, validate = FALSE)
Preprocess(gas_data = data.frame(date=as.Date(c("2024-06-20","2024-06-21","2024-06-22","2024-06-23","2024-06-24")), H2=c(80,80,80,80,82),CH4=c(90,95,100,105,107),C2H6=c(80,100,80,80,82),C2H4=c(30,30,31,29,31),C2H2=c(1,1,0,1,3),CO=c(800,850,860,850,852),CO2=c(900,950,940,940,942)),
           limit_data = data.frame(H2=c(80,200,40,20),CH4=c(90,150,30,10),C2H6=c(90,175,25,9),C2H4=c(50,100,20,7),C2H2=c(1,2,0.5,0.5),CO=c(900,1100,20,100),CO2=c(9000,12500,250,1000)),
           accuracy = 0.15,
           min_accuracy = 1,
           distribution_shape = "T",
           rho = matrix(rep(1,49), ncol=7),
           diagnosis = "DT",
           n = 1000,
           flex = TRUE, silent = FALSE, verbose = TRUE, validate = TRUE)
}
