% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_plot.R
\name{Sample_Plot}
\alias{Sample_Plot}
\title{Sample_Plot: Plotting outputs from Sample function for validation.}
\usage{
Sample_Plot(y_N, PP)
}
\arguments{
\item{y_N}{Numeric Array: Values representing generated samples for the Monte Carlo based on the statistical description provided. Shaped: Days x Gases x Samples.}

\item{PP}{List: Outputs from Preprocess function.}
}
\description{
This is internal function to plot generated samples for validation.
}
\examples{
# Not recreating required inputs to run this example
# Sample_Plot(y_1 = y_N[1,,],
#             PP = preprocessed)
}
